backend/requirements.txt
fastapi==0.111.1
uvicorn==0.30.5
python-multipart==0.0.9
torch==2.4.0
torchvision==0.19.0
Pillow==10.4.0
numpy==1.26.4
efficientnet-pytorch==0.7.1